Move restore helper to its own package
